7901085: VMSupport.guessAlignment calls System.gc, deteriorates performance when used as a library
Revisit the guessAlignment implementation, get rid of System.gc.

+++ jol-core/src/main/java/org/openjdk/jol/util/MathUtil.java
@@ -45,9 +45,9 @@ public class MathUtil {
         return min;
     }
 
-    public static int gcd(int a, int b) {
+    public static long gcd(long a, long b) {
         while (b > 0) {
-            int temp = b;
+            long temp = b;
             b = a % b;
             a = temp;
         }
